Perform ModuleOutDir cleanup

+++ build.psake.ps1
@@ -89,10 +89,11 @@ Task Init -requiredVariables OutDir {
     }
 }
 
-Task Clean -depends Init -requiredVariables OutDir {
+Task Clean -depends Init -requiredVariables OutDir, ModuleOutDir {
     # Maybe a bit paranoid but this task nuked \ on my laptop. Good thing I was not running as admin.
-    if ($OutDir.Length -gt 3) {
+    if ($OutDir.Length -gt 3 -and $ModuleOutDir.Length -gt 3) {
         Get-ChildItem $OutDir | Remove-Item -Recurse -Force -Verbose:$VerbosePreference
+        Get-ChildItem $ModuleOutDir | Remove-Item -Recurse -Force -Verbose:$VerbosePreference
     }
     else {
         Write-Verbose "$($psake.context.currentTaskName) - `$OutDir '$OutDir' must be longer than 3 characters."
@@ -110,7 +111,7 @@ Task CoreStageFiles -requiredVariables ModuleOutDir, SrcRootDir {
         Write-Verbose "$($psake.context.currentTaskName) - directory already exists '$ModuleOutDir'."
     }
 
-    Copy-Item -Path $SrcRootDir\* -Destination $ModuleOutDir -Recurse -Exclude $Exclude -Verbose:$VerbosePreference
+    Copy-Item -Path $SrcRootDir\* -Destination $ModuleOutDir -Recurse -Exclude $Exclude -Verbose:$VerbosePreference -Force
 }
 
 Task Build -depends Init, Clean, BeforeBuild, StageFiles, Analyze, Sign, AfterBuild {
